@@ -6,10 +6,10 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.1.1",
     "@testing-library/user-event": "^13.5.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "5.0.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^2.1.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -35,21 +35,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  }
-  
-}
-
- "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "express-async-handler": "^1.1.4",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.10.7"
-  },
-  "devDependencies": {
-    "eslint": "^7.9.0",
-    "eslint-plugin-react": "^7.21.2",
-    "nodemon": "^2.0.4"
   }
 }
